libpod: Move miscellaneous file handlling to container_internal_common.go
[NO NEW TESTS NEEDED]
